+++ cotests/cases/runner/__init__.py
@@ -1,0 +1,1 @@
+from .group import RootGroupRunner
